backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.66.3
Werkzeug==2.3.7
gunicorn==21.2.0
python-multipart==0.0.6 